refactor: (#281) Request → Data 변경
