+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/schedule/dto/request/ChangeIndividualScheduleData.kt
@@ -1,4 +1,4 @@
-package team.comit.simtong.domain.schedule.dto
+package team.comit.simtong.domain.schedule.dto.request
 
 import java.time.LocalDate
 import java.time.LocalTime
@@ -6,13 +6,13 @@ import java.util.UUID
 
 /**
  *
- * 개인 일정 변경 요청 정보를 전달하는 ChangeIndividualScheduleRequest
+ * 개인 일정 변경 요청 정보를 전달하는 ChangeIndividualScheduleData
  *
  * @author Chokyunghyeon
  * @date 2022/11/27
- * @version 1.0.0
+ * @version 1.2.5
  **/
-data class ChangeIndividualScheduleRequest(
+data class ChangeIndividualScheduleData(
 
     val scheduleId: UUID,
